feat: implement dynamic product filter search with specification, pagination and sorting

+++ src/main/java/com/florastore/web_ban_hoa/dto/ProductDTO.java
@@ -3,7 +3,7 @@ package com.florastore.web_ban_hoa.dto;
 public record ProductDTO(
         Long id,
         String name,
-        Long price,
+        BigDecimal price,
         String description,
         String imageUrl,
         String categoryName
