+++ src/main/java/com/florastore/web_ban_hoa/specification/ProductSpecification.java
@@ -2,6 +2,7 @@ package com.florastore.web_ban_hoa.specification;
 
 import com.florastore.web_ban_hoa.domain.Product;
 import org.springframework.data.jpa.domain.Specification;
+import java.math.BigDecimal;
 
 public final class ProductSpecification {
 
@@ -19,24 +20,30 @@ public final class ProductSpecification {
     }
 
     public static Specification<Product> hasPriceBetween(Long minPrice, Long maxPrice) {
-        return (root, query, cb) -> {
-            if (minPrice == null && maxPrice == null) {
-                return null;
-            }
+    return (root, query, cb) -> {
+        if (minPrice == null && maxPrice == null) {
+            return null;
+        }
 
-            if (minPrice != null && maxPrice != null) {
-                long lowerBound = Math.min(minPrice, maxPrice);
-                long upperBound = Math.max(minPrice, maxPrice);
-                return cb.between(root.get("price"), lowerBound, upperBound);
-            }
+        if (minPrice != null && maxPrice != null) {
+            BigDecimal lowerBound = BigDecimal.valueOf(Math.min(minPrice, maxPrice));
+            BigDecimal upperBound = BigDecimal.valueOf(Math.max(minPrice, maxPrice));
+            return cb.between(root.get("price"), lowerBound, upperBound);
+        }
 
-            if (minPrice != null) {
-                return cb.greaterThanOrEqualTo(root.get("price"), minPrice);
-            }
+        if (minPrice != null) {
+            return cb.greaterThanOrEqualTo(
+                    root.get("price"),
+                    BigDecimal.valueOf(minPrice)
+            );
+        }
 
-            return cb.lessThanOrEqualTo(root.get("price"), maxPrice);
-        };
-    }
+        return cb.lessThanOrEqualTo(
+                root.get("price"),
+                BigDecimal.valueOf(maxPrice)
+        );
+    };
+}
 
     public static Specification<Product> hasCategory(Long categoryId) {
         return (root, query, cb) -> {
